added validation to table actions

+++ softawere/table-actions/drop_table.sh
@@ -1,9 +1,15 @@
 #!/usr/bin/bash
 
 read -p "Enter the name of the table you want to drop: " tablename
-rm "${tablename}" "${tablename}_metadata" 2>>./.error.log
-if [[ $? == 0 ]]; then
-    echo "Table dropped successfully."
+
+# Check if the table and its metadata file exist
+if [[ -f "${tablename}" && -f "${tablename}_metadata" ]]; then
+    rm "${tablename}" "${tablename}_metadata" 2>>./.error.log
+    if [[ $? == 0 ]]; then
+        echo "Table dropped successfully."
+    else
+        echo "Error dropping table $tablename."
+    fi
 else
-    echo "Error dropping table $tablename."
+    echo "Table $tablename does not exist."
 fi
